fix clang-tidy warnings for a batch of files

+++ dbms/src/Flash/Coprocessor/DAGDriver.cpp
@@ -152,7 +152,7 @@ try
         {
             // Under some test cases, there may be dag response whose size is bigger than INT_MAX, and GRPC can not limit it.
             // Throw exception to prevent receiver from getting wrong response.
-            if (p_stream->getProfileInfo().bytes > std::numeric_limits<int>::max())
+            if (p_stream->getProfileInfo().bytes > static_cast<size_t>(std::numeric_limits<int>::max()))
                 throw TiFlashException("DAG response is too big, please check config about region size or region merge scheduler",
                                        Errors::Coprocessor::Internal);
         }
